pyproject.toml
[tool.poetry]
name = "course-project"
version = "0.1.0"
description = ""
authors = ["Constantine Putilov <putilov.constantine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
django-rest-framework = "^0.1.0"
envparse = "^0.2.0"
django-phonenumber-field = "^7.3.0"
pillow = "^10.2.0"
drf-spectacular = "^0.27.1"
psycopg2 = "^2.9.9"
phonenumbers = "^8.13.30"
django-polymorphic = "^3.1.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
stripe = "^8.6.0"

[tool.poetry.group.test.dependencies]
pytest-factoryboy = "^2.6.0"
pytest = "^8.0.2"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
